move secrets_importer and replacer modules into context package
