+++ tests/dataverk/context/test_secrets_importer.py
@@ -3,7 +3,7 @@
 # =================
 import unittest
 from dataverk.utils.windows_safe_tempdir import WindowsSafeTempDirectory
-from dataverk.context.secset_replacer.secrets_importer import get_secrets_importer, APISecretsImporter, FileSecretsImporter
+from dataverk.context.secrets_importer import get_secrets_importer, APISecretsImporter, FileSecretsImporter
 from pathlib import Path
 
 # Common input parameters
